Use overall results from latest event as Home
Adds the option to display the latest events Overall results as Home (index.js).

You set it up in initialState.js with
useResultsForHome: true,

+++ src/output/html.js
@@ -212,7 +212,7 @@ const writeStandingsHTML = (division, type, links) => {
     `./${outputPath}/website/${division.divisionName}-${type}-standings.html`,
     out
   );
-  if (!fs.existsSync(`./${outputPath}/website/index.html`)) {
+  if (!fs.existsSync(`./${outputPath}/website/index.html`) && leagueRef.league.useStandingsForHome) {
     fs.writeFileSync(`./${outputPath}/website/index.html`, out);
   }
 };
@@ -409,6 +409,9 @@ const writeDriverResultsHTML = (event, division, links, eventIndex) => {
     `./${outputPath}/website/${division.divisionName}-${eventIndex}-driver-results.html`,
     out
   );
+  if (division.divisionName == "overall" && leagueRef.league.useResultsForHome) {
+    fs.writeFileSync(`./${outputPath}/website/index.html`, out);
+  }
 };
 
 const addLinks = (links, name, type, displayName) => {
